Move memory-related functions to separate file
